printer-usuage-optimization-main/requirements.txt
sqlalchemy>=1.4
mysql-connector-python>=8.0
requests>=2.25
tqdm
streamlit
scikit-learn
pandas
numpy